+++ modules/matterhorn-common/src/main/java/org/opencastproject/job/api/JobImpl.java
@@ -97,7 +97,7 @@ public class JobImpl implements Job {
    */
   @XmlAttribute
   @Override
-  public int getVersion() {
+  public long getVersion() {
     return version;
   }
